Добавлены плейсхолдеры состояния экрана поиска Выстроена структура папок presentation слоя
