+++ app/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
         android:theme="@style/Theme.Subzy"
         tools:targetApi="31">
         <activity
-            android:name=".presentation.ui.root.RootActivity"
+            android:name=".presentation.root.RootActivity"
             android:screenOrientation="portrait"
             android:exported="true"
             android:windowSoftInputMode="adjustNothing"
