+++ app/src/main/java/com/anshok/subzy/domain/api/SearchRepository.kt
@@ -1,8 +1,8 @@
 package com.anshok.subzy.domain.api
 
 import com.anshok.subzy.domain.model.Logo
-import com.anshok.subzy.util.ResourceLogo
+import com.anshok.subzy.util.ResponseData
 
 interface SearchRepository {
-    suspend fun searchCompany(query: String): ResourceLogo<Logo>
+    suspend fun searchCompany(query: String): ResponseData<Logo>
 }
