+++ app/src/main/java/com/anshok/subzy/domain/impl/SearchInteractorImpl.kt
@@ -3,13 +3,13 @@ package com.anshok.subzy.domain.impl
 import com.anshok.subzy.domain.api.SearchInteractor
 import com.anshok.subzy.domain.api.SearchRepository
 import com.anshok.subzy.domain.model.Logo
-import com.anshok.subzy.util.ResourceLogo
+import com.anshok.subzy.util.ResponseData
 
 class SearchInteractorImpl(
     private val repository: SearchRepository
 ) : SearchInteractor {
 
-    override suspend fun searchCompany(query: String): ResourceLogo<Logo> {
+    override suspend fun searchCompany(query: String): ResponseData<Logo> {
         return repository.searchCompany(query)
     }
 }
