+++ app/src/main/java/com/anshok/subzy/presentation/add_sub/create/AddSubCreateFragment.kt
@@ -1,4 +1,4 @@
-package com.anshok.subzy.presentation.ui
+package com.anshok.subzy.presentation.add_sub.create
 
 import android.app.DatePickerDialog
 import android.os.Bundle
@@ -11,11 +11,11 @@ import by.kirich1409.viewbindingdelegate.CreateMethod
 import by.kirich1409.viewbindingdelegate.viewBinding
 import com.anshok.subzy.R
 import com.anshok.subzy.databinding.FragmentAddSubCreateBinding
-import com.anshok.subzy.presentation.ui.bottomSheet.CategoryBottomSheetFragment
-import com.anshok.subzy.presentation.ui.bottomSheet.CommentBottomSheet
-import com.anshok.subzy.presentation.ui.bottomSheet.DescriptionBottomSheet
-import com.anshok.subzy.presentation.ui.bottomSheet.NotificationReminderBottomSheet
-import com.anshok.subzy.presentation.ui.bottomSheet.PaymentPeriodBottomSheet
+import com.anshok.subzy.presentation.bottomSheetCreateSub.CategoryBottomSheetFragment
+import com.anshok.subzy.presentation.bottomSheetCreateSub.CommentBottomSheet
+import com.anshok.subzy.presentation.bottomSheetCreateSub.DescriptionBottomSheet
+import com.anshok.subzy.presentation.bottomSheetCreateSub.NotificationReminderBottomSheet
+import com.anshok.subzy.presentation.bottomSheetCreateSub.PaymentPeriodBottomSheet
 import java.text.SimpleDateFormat
 import java.util.Calendar
 import java.util.Locale
