+++ app/src/main/java/com/anshok/subzy/presentation/settings/SettingsFragment.kt
@@ -1,4 +1,4 @@
-package com.anshok.subzy.presentation.ui
+package com.anshok.subzy.presentation.settings
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -8,11 +8,11 @@ import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
 import by.kirich1409.viewbindingdelegate.CreateMethod
 import by.kirich1409.viewbindingdelegate.viewBinding
-import com.anshok.subzy.databinding.FragmentDetailsSubBinding
+import com.anshok.subzy.databinding.FragmentSettingsBinding
 
-class DetailsSubFragment : Fragment() {
+class SettingsFragment : Fragment() {
 
-    private val binding: FragmentDetailsSubBinding by viewBinding(CreateMethod.INFLATE)
+    private val binding: FragmentSettingsBinding by viewBinding(CreateMethod.INFLATE)
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -25,9 +25,9 @@ class DetailsSubFragment : Fragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        // Обработка нажатия кнопки "Назад"
+        // Обработка нажатия на кнопку "Назад"
         binding.backButton.setOnClickListener {
-            findNavController().navigateUp()
+            findNavController().navigateUp()  // Возвращение на предыдущий экран
         }
     }
 }
