+++ app/src/main/res/layout/activity_root.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.ui.root.RootActivity">
+    tools:context=".presentation.root.RootActivity">
 
     <!-- Основной контейнер для фрагментов -->
     <androidx.fragment.app.FragmentContainerView
