+++ app/src/main/res/layout/fragment_add_sub_create.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/Gray_100"
-    tools:context=".presentation.ui.AddSubCreateFragment">
+    tools:context=".presentation.add_sub.create.AddSubCreateFragment">
 
     <!-- Верхняя карточка с иконкой и названием -->
     <com.google.android.material.card.MaterialCardView
