+++ app/src/main/res/layout/fragment_calendar.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.ui.CalendarFragment">
+    tools:context=".presentation.calendar.CalendarFragment">
 
     <TextView
         android:id="@+id/text_calendar"
