+++ app/src/main/res/layout/fragment_home.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.ui.HomeFragment">
+    tools:context=".presentation.home.HomeFragment">
 
     <com.google.android.material.card.MaterialCardView
         android:id="@+id/upcard"
