+++ app/src/main/res/layout/fragment_home_rv_your_sub.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.ui.HomeRvYourSubFragment">
+    tools:context=".presentation.home.HomeRvSubFragment">
 
     <!-- RecyclerView для списка подписок -->
     <androidx.recyclerview.widget.RecyclerView
