+++ app/src/main/res/layout/fragment_spend.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.ui.SpendFragment">
+    tools:context=".presentation.spend.SpendFragment">
 
     <TextView
         android:id="@+id/text_calendar"
