+++ app/src/main/res/layout/fragment_wallet.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.ui.WalletFragment">
+    tools:context=".presentation.wallet.WalletFragment">
 
     <TextView
         android:id="@+id/text_wallet"
